Improve the calabash-ios script's command line usage help
The syntax of the command line usage help is non-standard. This commit changes the syntax to follow the convention used by "git --help". Namely:

Arguments in angled brackets are to be supplied by the user (all others are literals)
Arguments in square brackets are optional (all others are required)
Arguments in curly braces and separated by a pipe are mutually exclusive