should relaunch and wait for directories to be recreated
This is very tricky and I am not satisified with my solution.

The problem is that if you reset the content and settings on the
simulator and immediately try to launch a cucumber job via
instruments, you will find that accessibility is not enabled because
the accessibility plist will not have been created.

This is a partial solution - relaunch and wait for the accessibility
plist to be created;  ~3.0 - ~5.0 seconds.

I think the complete solution is to relaunch _each_ available
simulator and wait for _all_ accessibility plists to appear.  It will
look strange, but it will be 'safer'.