Run rake-build-server last in CI
There is an order of operations conflict. The objc
required libs required by the gem need to be
present in order for the gem to be installed.

ATM the server make dylibs step cannot pass because
it requires xcspecs to be injected into the Xcode.app
bundle.

The build-server tasks deletes the stub (fake) libraries
installed in a previous CI build step.