CLI: remove 'update' command
This was a tool that was added as one-shot back in 2012 to transition
user's hooks.  Some tool like this might need to exist, but the
current tool is out of date and irrelvant.  The word 'update' is also
overloaded because it usually means, 'update your server version or
update your gem version'